CRM/requirements.txt
pip
tokenizers 
sqlite-fts4                   
sqlite-utils                  
sqlparse
setuptools 
requests 
numpy
multidict
httplib2 
google-cloud-core 
google-api-core              
google-api-python-client
djoser 
Django                        
django-allauth                
django-cors-headers           
django-crispy-forms           
django-environ                
django-filter                 
django-heroku                 
django-phonenumber-field
django-shortcuts             
django-templated-mail        
djangorestframework 
decorator
